Split import code and django-admin handling, move code to batch/incwo
